Altered shader lens logic to use maxDimension instead of minDimension
I realized with the stickyHeader and grid scaffold demos that something was wrong with the effect when using it on non-square shapes where one dimension was much larger than the other. The lens distortion would stretch way too strongly along the larger dimension creating a bizarre look.

I believe this may now be solved as the lens formula was originally written where it always expected a square shape, and when I made adjustments to support multiple cornerRadii, we needed to calculate shapes and edge effects by the smaller dimension. However given the m2 and lens logic, if say one dimension was four times larger than the other, the sdf would stretch four times stronger across the larger coordinate.

I decided to create this as a PR because I want to test that the Roborazzi build.yml logic works in the first place for this, but I also want to inspect the screenshot tests in the PR. I know it will fail, but I want to see how the github actions handle all of this. Also I had previously accounted for the distortion by setting a larger refraction (0.4 instead of 0.25) and creating a smaller curve (0.15 instead of 0.25). I think now I could revert at least the refraction and we should see a better lens effect for these rectangle shapes. Also importantly the drag and many liquid node demos should see no difference as they're squares, although maybe the slider box might be altered a bit since it isn't a square, but close.