Added boundsInRoot to the dirty tracker
- This may be a compose bug, but I discovered when adjusting the LiquidControls to use AnimatedVisibility in a more correct manner (the fadeIn/Out never ran, the liquid node should be on the LazyColumn's modifier and not AnimatedVisibility), that the enter animations wouldn't draw the liquid effect.

After some debugging, I discovered that when animating the view back in, the onGloballyPositioned would trigger with correct sizes and screen positions, but the `coordinates.boundsInRoot()` was returning with an empty rect. Only on one of the next frames would it update with correct bounds. However since the existing tracker assumed size and positionOnScreen would be in sync with boundsInRoot, we weren't calling `invalidateDraw()`.

- Also added some adaptive layout logic back as this helped with debugging.